% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcount.R
\name{kcount}
\alias{kcount}
\title{\code{kcount}
Counting kmer frequency for a given genome or pseudo genome.}
\usage{
kcount(fa, feature, kmers = 3:7)
}
\arguments{
\item{fa}{Path for the reference fasta file. [string or DNAStringSet/DNAString object]}

\item{feature}{A data.frame object (BED like, but all 1-based coordinates).
[data.frame, 4 required columns: chr, start, end, geneid]}

\item{kmers}{A vector to indicate kmers to count. Note, k<=15. [vector, default=3:7].}
}
\value{
A data.frame of kmer as rows x geneid as columns. [data.frame].
}
\description{
\code{kcount}
Counting kmer frequency for a given genome or pseudo genome.
}
\examples{

feature <- data.frame(chr=1, start=seq(from=1, to=10000001, by=1000),
                      end=seq(from=1000, to=11000000, by=1000)[1:10001],
                      geneid=paste0("g", 1:10001))
kcount <- function(fa, feature, kmers=3:7)

}

