% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoRef.R
\name{pseudoRef}
\alias{pseudoRef}
\title{\code{pseudoRef}
Make a pseudo reference genome.}
\usage{
pseudoRef(fa, snpdt, sidx = 5:ncol(snpdt), arules = NULL, outdir)
}
\arguments{
\item{fa}{Path for the reference fasta file. [string or DNAStringSet/DNAString oject]}

\item{snpdt}{A data.table object with heterozygote SNPs coded with IUPAC ambiguity codes.
[data.table, 4 required columns: chr, pos, ref, alt, (sample1, ..., sampleN)]}

\item{sidx}{A vector to indicate the sample columns. [vector, default=5:ncol(snpdt)].}

\item{arules}{Additional nucleotide substitution rules defined by users.
[data.frame, 2 required columns: from, to, default=NULL]
For example, arules <- data.frame(from=c("M", "Y", "R", "K"), to=c("C", "C", "G", "T")).}

\item{outdir}{Output directory. Sample specific sub-folders will be created. [string]}
}
\value{
A list of summary statistics of subsituted nucleotides. [list].
}
\description{
\code{pseudoRef}
Make a pseudo reference genome.
}
\examples{
# First of all, use BCFtools to convert VCF into IUPAC coded data.table:

# bcftools view JRI20_filtered_snps_annot.bcf.gz -m2 -M2 -v snps -Oz -o JRI20_bi_snps_annot.vcf.gz
# bcftools query -f '\%CHROM\\t\%POS\\t\%REF\\t\%ALT[\\t\%IUPACGT]\\n' JRI20_bi_snps_annot.vcf.gz > JRI20_bi_snps_annot.txt
# bcftools query -f 'chr\\tpos\\tref\\talt[\\t\%SAMPLE]\\n' JRI20_bi_snps_annot.vcf.gz > JRI20_bi_snps_annot.header

arules <- data.frame(from=c("M", "Y", "R", "K"), to=c("C", "C", "G", "T"))
res <- pseudoRef(fa, snpdt, sidx=5:24, arules, outdir)

}

