% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoRef.R
\name{pseudoRef}
\alias{pseudoRef}
\title{\code{pseudoRef}
Make a pseudo reference genome.}
\usage{
pseudoRef(fa, snpdt, sampleidx = 5:24, sub_rules)
}
\arguments{
\item{fa}{Path for the fasta file. [string]}

\item{snpdt}{A data.table object with heterozygote IUPAC SNP codes. [data.table]}

\item{sampleidx}{A vector to indicate the sample columns. [vector].}

\item{sub_rules}{Nucleotide substitution rules defined by users. [data.frame]
For example, sub_rules <- data.frame(from=c("M", "Y", "R", "K"), to=c("C", "C", "G", "T")).}

\item{dir}{Output directory. Sample specific sub-folders will be created. [string]}
}
\value{
Fasta files in specified directory under sample-specific sub-folders. [fasta].
}
\description{
\code{pseudoRef}
Make a pseudo reference genome.
}
\examples{
sub_rules <- data.frame(from=c("M", "Y", "R", "K"), to=c("C", "C", "G", "T"))

features <- c("C methylated in CHH context")
get_file2tab(files, features, replace=T )

}

